Create folder for tests, move file
